tf-k8s/requirements.txt
# TF-K8s Validation Framework Dependencies

# Core dependencies
pyyaml>=6.0,<7.0
click>=8.1.3,<9.0
jsonschema>=4.17.3,<5.0
jinja2>=3.1.2,<4.0
pytest>=7.3.1,<8.0
pytest-timeout>=2.1.0,<3.0
requests>=2.30.0,<3.0
kubernetes>=27.2.0,<28.0

# Validation tools
yamllint>=1.31.0,<2.0
markdownlint-cli>=0.35.0,<1.0

# Security scanning
grype>=0.65.0,<1.0
syft>=0.84.1,<1.0

# OTTL helpers
opentelemetry-collector-contrib>=0.84.0,<1.0

# Visualization
matplotlib>=3.7.1,<4.0
pandas>=2.0.2,<3.0
seaborn>=0.12.2,<1.0
